Вторая сдача промежуточной работы
