+++ src/components/app-header/app-header.module.css
@@ -2,6 +2,7 @@
     box-sizing: border-box;
     display: flex;
     justify-content: center;
+    background: #1C1C21;
 }
 
 .container {
