+++ src/components/bc-bun/bc-bun.module.css
@@ -1,0 +1,20 @@
+.bun {
+    display: flex;
+    align-items: center;
+    background: #1C1C21;
+}
+
+.image {
+    width: 80px;
+    height: 40px;
+}
+
+.bunTop {
+    composes: bun;
+    border-radius: 88px 88px 40px 40px;
+}
+
+.bunBottom {
+    composes: bun;
+    border-radius: 40px 40px 88px 88px;
+}
