+++ src/components/bc-ingredient/bc-ingredient.module.css
@@ -1,0 +1,23 @@
+.ingredientBlock {
+    display: flex;
+    align-items: center;
+}
+
+.ingredient {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+    border-radius: 40px;
+    background: #1C1C21;
+}
+
+
+.image {
+    width: 80px;
+    height: 40px;
+}
+
+.name {
+    width: 254px;
+}
