+++ src/components/burger-constructor-confirm/burger-constructor-confirm.module.css
@@ -1,0 +1,10 @@
+.register {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+}
+
+.cost {
+    display: flex;
+    align-items: center;
+}
