+++ src/components/burger-constructor-list/burger-constructor-list.module.css
@@ -1,0 +1,5 @@
+.ingredients {
+    height: 464px;
+    overflow: scroll;
+    overflow-x: hidden;
+}
