+++ src/components/burger-constructor/burger-constructor.module.css
@@ -36,20 +36,3 @@
     width: 80px;
     height: 40px;
 }
-
-.ingredients {
-    height: 464px;
-    overflow: scroll;
-    overflow-x: hidden;
-}
-
-.register {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-}
-
-.cost {
-    display: flex;
-    align-items: center;
-}
