+++ src/components/ingredient-cards/ingredient-cards.module.css
@@ -1,0 +1,6 @@
+.cards {
+    display: flex;
+    width: 576px;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
